Move path utilities into path crate
