+++ crates/language/Cargo.toml
@@ -31,6 +31,7 @@ imara-diff = { workspace = true }
 language_core = { workspace = true }
 log = { workspace = true }
 parking_lot = { workspace = true }
+path = { workspace = true }
 settings = { workspace = true }
 streaming-iterator = { workspace = true }
 sum_tree = { workspace = true }
