+++ crates/path/Cargo.toml
@@ -8,9 +8,21 @@ publish.workspace = true
 [lints]
 workspace = true
 
+[features]
+test-support = []
+
 [lib]
 name = "path"
 path = "src/path.rs"
 
 [dependencies]
+anyhow = { workspace = true }
 dirs = { workspace = true }
+dunce = { workspace = true }
+serde = { workspace = true }
+
+[target.'cfg(target_os = "windows")'.dependencies]
+tendril = { workspace = true }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
