+++ crates/request_editor/Cargo.toml
@@ -42,6 +42,7 @@ gpui = { workspace = true, features = ["test-support"] }
 http_client = { workspace = true, features = ["test-support"] }
 indoc = { workspace = true }
 parking_lot = { workspace = true }
+path = { workspace = true, features = ["test-support"] }
 project = { workspace = true, features = ["test-support"] }
 serde_json = { workspace = true }
 settings = { workspace = true, features = ["test-support"] }
