+++ crates/sql/Cargo.toml
@@ -18,8 +18,8 @@ indoc = { workspace = true }
 libsqlite3-sys = { workspace = true }
 num-traits = { workspace = true }
 parking_lot = { workspace = true }
+path = { workspace = true }
 thread_local = { workspace = true }
-util = { workspace = true }
 uuid = { workspace = true }
 
 [dev-dependencies]
