+++ crates/util/Cargo.toml
@@ -16,15 +16,9 @@ path = "src/util.rs"
 test-support = []
 
 [dependencies]
-anyhow = { workspace = true }
-dunce = { workspace = true }
 log = { workspace = true }
 rust-embed = { workspace = true }
 serde = { workspace = true }
 
 [target.'cfg(target_os = "windows")'.dependencies]
 smol = { workspace = true }
-tendril = { workspace = true }
-
-[dev-dependencies]
-pretty_assertions = { workspace = true }
