+++ crates/workspace/Cargo.toml
@@ -39,6 +39,7 @@ language = { workspace = true }
 log = { workspace = true }
 metadata = { workspace = true }
 parking_lot = { workspace = true }
+path = { workspace = true }
 project = { workspace = true }
 session = { workspace = true }
 serde = { workspace = true }
@@ -57,6 +58,7 @@ db = { workspace = true, features = ["test-support"] }
 fs = { workspace = true, features = ["test-support"] }
 gpui = { workspace = true, features = ["test-support"] }
 http_client = { workspace = true, features = ["test-support"] }
+path = { workspace = true, features = ["test-support"] }
 project = { workspace = true, features = ["test-support"] }
 session = { workspace = true, features = ["test-support"] }
 settings = { workspace = true, features = ["test-support"] }
