+++ crates/worktree/Cargo.toml
@@ -27,6 +27,7 @@ futures = { workspace = true }
 gpui = { workspace = true }
 language = { workspace = true }
 log = { workspace = true }
+path = { workspace = true }
 serde = { workspace = true }
 settings = { workspace = true }
 smallvec = { workspace = true }
@@ -42,6 +43,7 @@ fs = { workspace = true, features = ["test-support"] }
 gpui = { workspace = true, features = ["test-support"] }
 indoc = { workspace = true }
 parking_lot = { workspace = true }
+path = { workspace = true, features = ["test-support"] }
 pretty_assertions = { workspace = true }
 serde_json = { workspace = true }
 util_macros = { workspace = true }
